requirements.txt
fastapi
uvicorn
python-multipart
gtts
python-dotenv
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
